update rabbitmq parser, add redis support cluster conf, jemalloc as allocator, add tcp metrics from aggregator, update pushgateway metrics, remove old files

+++ src/parsers/http_proto.h
@@ -10,7 +10,7 @@
 typedef struct http_reply_data
 {
 	int http_version;
-	int http_code;
+	uint16_t http_code;
 	char *mesg;
 	char *headers;
 	char *uri;
